+++ packages/core/src/transport/index.ts
@@ -1,0 +1,2 @@
+export { HttpRequest } from './httpRequest'
+export { Batch } from './batch'
